use our new watt-pickers
