+++ libs/watt/src/lib/components/picker/datepicker/+storybook/watt-datepicker-reactive-forms.stories.ts
@@ -27,7 +27,6 @@ import { WattRangeValidators } from '../../shared/validators';
 
 import { WattDateChipComponent } from '../watt-date-chip.component';
 import { WattDateRangeChipComponent } from '../watt-date-range-chip.component';
-import { WattDatepickerV2Component } from '../watt-datepicker-v2.component';
 
 export const initialValueSingle = '2022-09-02T22:00:00.000Z';
 export const initialValueRangeStart = initialValueSingle;
@@ -52,7 +51,6 @@ export default {
         WattDateChipComponent,
         WattDateRangeChipComponent,
         WattFieldErrorComponent,
-        WattDatepickerV2Component,
       ],
     }),
   ],
@@ -67,22 +65,22 @@ export default {
 
 const template = `
 
-  <watt-datepicker-v2 label="Single date" [formControl]="exampleFormControlSingle">
+  <watt-datepicker label="Single date" [formControl]="exampleFormControlSingle">
     <watt-field-error *ngIf="exampleFormControlSingle?.errors?.required">
       Date is required
     </watt-field-error>
-  </watt-datepicker-v2>
+  </watt-datepicker>
 
 <p>Value: <code>{{ exampleFormControlSingle.value | json }}</code></p>
 <p *ngIf="withValidations">Errors: <code>{{ exampleFormControlSingle?.errors | json }}</code></p>
 
 <br />
 
-<watt-datepicker-v2 label="Date range" [formControl]="exampleFormControlRange" [range]="true">
+<watt-datepicker label="Date range" [formControl]="exampleFormControlRange" [range]="true">
   <watt-field-error *ngIf="exampleFormControlRange?.errors?.rangeRequired">
     Date range is required
   </watt-field-error>
-</watt-datepicker-v2>
+</watt-datepicker>
 
 <p>Selected range: <code data-testid="rangeValue">{{ exampleFormControlRange.value | json }}</code></p>
 <p *ngIf="withValidations">Errors: <code>{{ exampleFormControlRange?.errors | json }}</code></p>
