+++ libs/watt/src/lib/components/picker/datepicker/index.ts
@@ -15,7 +15,6 @@
  * limitations under the License.
  */
 export { WattDatepickerComponent } from './watt-datepicker.component';
-export { WattDatepickerV2Component } from './watt-datepicker-v2.component';
 export { WattDateRangeChipComponent } from './watt-date-range-chip.component';
 export { WattDateChipComponent } from './watt-date-chip.component';
 export { danishTimeZoneIdentifier } from './watt-datepicker.component';
