--- libs/watt/src/lib/components/picker/datepicker/watt-datepicker-v2.component.html
@@ -1,123 +1,0 @@
-<!--
-@license
-Copyright 2020 Energinet DataHub A/S
-
-Licensed under the Apache License, Version 2.0 (the "License2");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-<ng-container *ngIf="range; else singleInputTemplate">
-  <watt-field [control]="control" [label]="label">
-    <mat-date-range-input
-      [disabled]="disabled"
-      [rangePicker]="rangeDatepicker"
-      [min]="min"
-      [max]="max"
-      (focusin)="onFocusIn()"
-      (focusout)="onFocusOut($event)"
-    >
-      <input
-        aria-label="start-date-input"
-        matStartDate
-        #startDateInput
-        [readOnly]="rangeMonthOnlyMode"
-        autocomplete="off"
-        spellcheck="false"
-        [hidden]="true"
-      />
-
-      <input
-        aria-label="end-date-input"
-        matEndDate
-        #endDateInput
-        [readOnly]="rangeMonthOnlyMode"
-        autocomplete="off"
-        spellcheck="false"
-        [hidden]="true"
-      />
-    </mat-date-range-input>
-    <input #actualInput class="mask-input" />
-    <watt-placeholder-mask
-      [primaryInputElement]="actualInput"
-      [mask]="rangeInputMask"
-      [placeholder]="rangePlaceholder"
-      (maskApplied)="rangeInputChanged($event)"
-    />
-
-    <watt-button
-      variant="icon"
-      icon="date"
-      [disabled]="disabled"
-      [attr.aria-pressed]="false"
-      (click)="rangeDatepicker.open()"
-      (focusin)="onFocusIn()"
-      (focusout)="onFocusOut($event)"
-    />
-
-    <ng-content />
-    <ng-content ngProjectAs="watt-field-hint" select="watt-field-hint" />
-    <ng-content ngProjectAs="watt-field-error" select="watt-field-error" />
-  </watt-field>
-  <mat-date-range-picker
-    [panelClass]="rangeMonthOnlyMode ? 'watt-datepicker-range__panel--month-only' : ''"
-    [startView]="rangeMonthOnlyMode ? 'multi-year' : 'month'"
-    [startAt]="startAt"
-    (monthSelected)="onMonthSelected($event)"
-    (closed)="rangePickerClosed()"
-    #rangeDatepicker
-  />
-</ng-container>
-
-<ng-template #singleInputTemplate>
-  <watt-field [control]="control" [label]="label">
-    <input
-      matInput
-      aria-label="date-input"
-      #dateInput
-      autocomplete="off"
-      spellcheck="false"
-      [disabled]="disabled"
-      [min]="min"
-      [max]="max"
-      [matDatepicker]="singleDatepicker"
-      (focusin)="onFocusIn()"
-      (focusout)="onFocusOut($event)"
-    />
-    <input #actualInput class="mask-input" />
-    <watt-placeholder-mask
-      [primaryInputElement]="actualInput"
-      [mask]="inputMask"
-      [placeholder]="datePlaceholder"
-      (maskApplied)="inputChanged($event)"
-    />
-
-    <watt-button
-      wattSuffix
-      variant="icon"
-      icon="date"
-      [disabled]="disabled"
-      (click)="singleDatepicker.open()"
-      (focusin)="onFocusIn()"
-      (focusout)="onFocusOut($event)"
-    />
-
-    <ng-content />
-    <ng-content ngProjectAs="watt-field-hint" select="watt-field-hint" />
-    <ng-content ngProjectAs="watt-field-error" select="watt-field-error" />
-  </watt-field>
-  <mat-datepicker
-    #singleDatepicker
-    panelClass="watt-datepicker-single__panel"
-    [dateClass]="dateClass"
-    [startAt]="startAt"
-    (closed)="datepickerClosed()"
-  />
-</ng-template>
