+++ libs/watt/src/lib/components/picker/datepicker/watt-datepicker.component.scss
@@ -17,7 +17,7 @@
 @use "../shared/picker" as *;
 
 watt-datepicker,
-watt-datepicker-v2 {
+watt-datepicker {
   @extend %datepicker;
 
   watt-button .mat-button {
