+++ libs/watt/src/lib/components/picker/timepicker/index.ts
@@ -15,4 +15,3 @@
  * limitations under the License.
  */
 export { WattTimepickerComponent } from './watt-timepicker.component';
-export { WattTimepickerV2Component } from './watt-timepicker-v2.component';
