--- libs/watt/src/lib/components/picker/timepicker/watt-timepicker-v2.component.html
@@ -1,112 +1,0 @@
-<!--
-@license
-Copyright 2020 Energinet DataHub A/S
-
-Licensed under the Apache License, Version 2.0 (the "License2");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-<ng-container *ngIf="!range; else rangeTemplate">
-  <watt-field [id]="id" [control]="control" [label]="label">
-    <input
-      matInput
-      aria-label="time-input"
-      #timeInput
-      autocomplete="off"
-      spellcheck="false"
-      [disabled]="disabled"
-      (focusin)="onFocusIn()"
-      (focusout)="onFocusOut($event)"
-      class="mask-input"
-    />
-    <watt-placeholder-mask
-      [primaryInputElement]="timeInput"
-      [mask]="inputMask"
-      [placeholder]="hoursMinutesPlaceholder"
-      (maskApplied)="inputChanged($event)"
-    />
-    <ng-content />
-    <ng-content ngProjectAs="watt-field-hint" select="watt-field-hint" />
-    <ng-content ngProjectAs="watt-field-error" select="watt-field-error" />
-  </watt-field>
-</ng-container>
-
-<ng-template #rangeTemplate>
-  <watt-field [id]="id" [control]="control" [label]="label">
-    <mat-date-range-input
-      cdkOverlayOrigin
-      #trigger="cdkOverlayOrigin"
-      [disabled]="disabled"
-      (focusin)="onFocusIn()"
-      (focusout)="onFocusOut($event)"
-    >
-      <input
-        aria-label="start-time-input"
-        matStartDate
-        #startTimeInput
-        autocomplete="off"
-        spellcheck="false"
-      />
-      <input
-        aria-label="end-time-input"
-        matEndDate
-        #endTimeInput
-        autocomplete="off"
-        spellcheck="false"
-      />
-    </mat-date-range-input>
-    <input #timeInput class="mask-input" />
-    <watt-placeholder-mask
-      [primaryInputElement]="timeInput"
-      [mask]="rangeInputMask"
-      [placeholder]="rangePlaceholder"
-      (maskApplied)="rangeInputChanged($event)"
-    />
-
-    <watt-button
-      variant="icon"
-      icon="time"
-      [disabled]="disabled"
-      [attr.aria-pressed]="sliderOpen"
-      (click)="toggleSlider()"
-      (focusin)="onFocusIn()"
-      (focusout)="onFocusOut($event)"
-    />
-
-    <ng-template
-      cdkConnectedOverlay
-      [cdkConnectedOverlayOrigin]="trigger"
-      [cdkConnectedOverlayOpen]="sliderOpen"
-      [cdkConnectedOverlayOffsetY]="12"
-    >
-      <div
-        [id]="sliderId"
-        role="dialog"
-        class="watt-timepicker-slider"
-        tabindex="-1"
-        (focusin)="onFocusIn()"
-        (focusout)="onFocusOut($event)"
-      >
-        <span class="watt-label">{{ sliderLabel }}</span>
-        <watt-slider
-          [min]="sliderSteps[0]"
-          [max]="sliderSteps[sliderSteps.length - 1]"
-          [step]="1"
-          [value]="sliderValue"
-          (valueChange)="sliderChange$.next($event)"
-        />
-      </div>
-    </ng-template>
-    <ng-content />
-    <ng-content ngProjectAs="watt-field-hint" select="watt-field-hint" />
-    <ng-content ngProjectAs="watt-field-error" select="watt-field-error" />
-  </watt-field>
-</ng-template>
