+++ libs/watt/src/lib/components/picker/timepicker/watt-timepicker.component.scss
@@ -17,7 +17,7 @@
 @use "../shared/picker" as *;
 
 watt-timepicker,
-watt-timepicker-v2 {
+watt-timepicker {
   @extend %datepicker;
 }
