+++ package.json
@@ -92,7 +92,6 @@
     "graphql": "16.8.1",
     "highlight.js": "^11.9.0",
     "include-media": "2.0.0",
-    "inputmask": "5.0.7",
     "jwt-decode": "3.1.2",
     "libphonenumber-js": "^1.10.54",
     "lottie-web": "5.12.2",
